# backend/auth-service/openapi/auth.yaml
openapi: 3.0.3
info:
  title: YUGEN Auth Service API
  description: Authentication and Authorization service for YUGEN Clothing E-Commerce Platform.
  version: 1.0.0
servers:
  - url: http://localhost:5001/auth
    description: Auth Service Direct
  - url: http://localhost:5000/auth
    description: API Gateway
paths:
  /register:
    post:
      summary: Register a new customer
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password]
              properties:
                firstName:
                  type: string
                  example: Jane
                lastName:
                  type: string
                  example: Doe
                email:
                  type: string
                  format: email
                  example: jane.doe@example.com
                password:
                  type: string
                  format: password
                  example: SecretPass123!
                phone:
                  type: string
                  example: "+919876543210"
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation error
        '409':
          description: Duplicate email conflict

  /login:
    post:
      summary: Login with credentials
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: jane.doe@example.com
                password:
                  type: string
                  format: password
                  example: SecretPass123!
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid email or password

  /refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      responses:
        '200':
          description: Access token refreshed
        '401':
          description: Invalid or expired refresh token

  /logout:
    post:
      summary: Logout user and revoke refresh token
      operationId: logoutUser
      responses:
        '200':
          description: Logged out successfully

  /me:
    get:
      summary: Get current authenticated user profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
        '401':
          description: Unauthorized

  /verify-email:
    post:
      summary: Verify user email with token
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Email verified successfully
        '400':
          description: Invalid or expired token

  /forgot-password:
    post:
      summary: Request password reset token
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset token issued

  /reset-password:
    post:
      summary: Reset password with token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid or expired reset token

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT